pyproject.toml
[project]
name = "touhou-card-bot"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "hikari>=2.0.0.dev111",
    "hikari-crescent>=0.3.6",
    "python-dotenv>=0.21.0",
    "hikari-flare @ git+https://github.com/Lunarmagpie/hikari-flare@feat/custom_dataclass",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
namespaces = true

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
strict = true
version = "3.11"
